git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/rejuvenate01@1488227 13f79535-47bb-0310-9956-ffa450edef68

+++ main/sc/source/ui/docshell/docsh.cxx
@@ -203,7 +203,7 @@ namespace
         static const struct
         {
             const char * mpFilterName;
-            unsigned mnFilterNameLen;
+            size_t mnFilterNameLen;
         } szMSFilterNames [] = 
         {
             { pFilterExcel4, strlen( pFilterExcel4 ) },
